requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3
redis==4.6.0
celery==5.3.1
boto3==1.28.85
botocore==1.31.85
python-dotenv==1.0.0
gunicorn==21.2.0
cryptography
psycopg2-binary
SQLAlchemy
Flask-SQLAlchemy
alembic
backoff==2.2.1
google-generativeai==0.7.1
